requirements.txt
aws-cdk-lib==2.79.1
aws-cdk.aws-lambda-python-alpha==2.79.1a0
constructs>=10.0.0,<11.0.0
